Combine all scripts into bundle
