+++ bookmarks/templates/bookmarks/form.html
@@ -118,7 +118,6 @@
 
   {# Replace tag input with auto-complete component #}
   <script src="{% static "bundle.js" %}"></script>
-  <script src="{% static "shared.js" %}"></script>
   <script type="application/javascript">
     /**
      * - Pre-fill title and description placeholders with metadata from website as soon as URL changes
