+++ bookmarks/templates/bookmarks/index.html
@@ -42,6 +42,4 @@
   </div>
 
   <script src="{% static "bundle.js" %}"></script>
-  <script src="{% static "shared.js" %}"></script>
-  <script src="{% static "bookmark_list.js" %}"></script>
 {% endblock %}
