+++ bookmarks/templates/bookmarks/shared.html
@@ -46,6 +46,4 @@
   </div>
 
   <script src="{% static "bundle.js" %}"></script>
-  <script src="{% static "shared.js" %}"></script>
-  <script src="{% static "bookmark_list.js" %}"></script>
 {% endblock %}
